requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-core==2.20.1
python-dotenv==1.0.1
firebase-admin==6.5.0
httpx==0.27.2
slowapi==0.1.9
pytest==8.3.3
pytest-asyncio==0.23.8
requests==2.32.3
google-cloud-firestore==2.16.0
pydantic-settings==2.3.4
